+++ engine/core/src/main/kotlin/io/canopy/engine/core/managers/SceneManager.kt
@@ -2,8 +2,8 @@ package io.canopy.engine.core.managers
 
 import kotlin.reflect.KClass
 import com.badlogic.gdx.math.Vector2
-import io.canopy.engine.core.nodes.core.Node
-import io.canopy.engine.core.nodes.core.TreeSystem
+import io.canopy.engine.core.nodes.Node
+import io.canopy.engine.core.nodes.TreeSystem
 import io.canopy.engine.core.reactive.asSignal
 import io.canopy.engine.core.reactive.event
 import io.canopy.engine.logging.EngineLogs
