+++ engine/core/src/main/kotlin/io/canopy/engine/core/nodes/CanopyDsl.kt
@@ -1,0 +1,4 @@
+package io.canopy.engine.core.nodes
+
+@DslMarker
+annotation class CanopyDsl
