+++ engine/core/src/main/kotlin/io/canopy/engine/core/nodes/types/empty/EmptyNode.kt
@@ -1,6 +1,6 @@
 package io.canopy.engine.core.nodes.types.empty
 
-import io.canopy.engine.core.nodes.core.Node
+import io.canopy.engine.core.nodes.Node
 
 /** Empty Node with no Behavior **/
 class EmptyNode(name: String, block: EmptyNode.() -> Unit = {}) : Node<EmptyNode>(name, block)
