+++ engine/core/src/test/kotlin/io/canopy/engine/core/nodes/NodeRefTests.kt
@@ -5,12 +5,9 @@ import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
 import io.canopy.engine.core.managers.ManagersRegistry
 import io.canopy.engine.core.managers.SceneManager
-import io.canopy.engine.core.nodes.core.Node
-import io.canopy.engine.core.nodes.core.NodeRef
-import io.canopy.engine.core.nodes.core.asSceneRoot
-import io.canopy.engine.core.nodes.core.behavior
-import io.canopy.engine.core.nodes.core.nodeRef
 import io.canopy.engine.core.nodes.types.empty.EmptyNode
+import io.canopy.engine.core.reactive.NodeRef
+import io.canopy.engine.core.reactive.nodeRef
 import org.junit.jupiter.api.BeforeAll
 
 class NodeRefTests {
