+++ engine/input/src/main/kotlin/io/canopy/engine/input/nodes/InputBehavior.kt
@@ -1,7 +1,7 @@
 package io.canopy.engine.input.nodes
 
-import io.canopy.engine.core.nodes.core.Behavior
-import io.canopy.engine.core.nodes.core.Node
+import io.canopy.engine.core.nodes.Behavior
+import io.canopy.engine.core.nodes.Node
 import io.canopy.engine.input.InputEvent
 
 /**
